refactor: 네이밍 변경 index.tsx 및 router 관련
