+++ src/app/main.tsx
@@ -3,7 +3,7 @@ import "@/app/styles/global.css";
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import App from "@/app";
+import App from "@/app/app";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
