--- src/pages/home/index.tsx
@@ -1,5 +1,0 @@
-import { HomePage } from "@/pages/home/ui/HomPage";
-
-export default function Home() {
-  return <HomePage />;
-}
