--- src/pages/home/model/index.ts
@@ -1,1 +1,0 @@
-export * from "@/pages/home/model/featureCards";
