--- src/pages/home/ui/index.ts
@@ -1,1 +1,0 @@
-export { HomeFeatureCards } from "@/pages/home/ui/HomeFeatureCards";
