+++ src/widgets/site-footer/index.ts
